Moved Utility to new class, MyUtils
