+++ core/src/miniventure/game/screen/GameScreen.java
@@ -3,6 +3,7 @@ package miniventure.game.screen;
 import java.util.HashMap;
 
 import miniventure.game.GameCore;
+import miniventure.game.MyUtils;
 import miniventure.game.item.Item;
 import miniventure.game.world.Level;
 import miniventure.game.world.entity.mob.Player;
@@ -159,7 +160,7 @@ public class GameScreen implements Screen {
 		if(heldItem != null) {
 			float x = camera.viewportWidth / 3;
 			batch.draw(heldItem.getTexture(), x, 5);
-			GameCore.writeOutlinedText(game.getFont(), batch, heldItem.getName(), x+heldItem.getTexture().getRegionWidth()+10, 5);
+			MyUtils.writeOutlinedText(game.getFont(), batch, heldItem.getName(), x+heldItem.getTexture().getRegionWidth()+10, 5);
 		}
 		
 		// a list of text to display in the upper left, for debug purposes
@@ -178,6 +179,6 @@ public class GameScreen implements Screen {
 		debugInfo.add("Looking at: " + (interactTile == null ? "Null" : interactTile.toLocString()));
 		
 		for(int i = 0; i < debugInfo.size; i++)
-			GameCore.writeOutlinedText(game.getFont(), batch, debugInfo.get(i), 0, uiCamera.viewportHeight-5-15*i);
+			MyUtils.writeOutlinedText(game.getFont(), batch, debugInfo.get(i), 0, uiCamera.viewportHeight-5-15*i);
 	}
 }
